.history/requirements_20241023165715.txt
transformers
diffusers[torch]
telegram-bot